Major overhaul to calculator
* The smallest successful formula for each possible roll combination has
been precalculated for every spell level and stored as binaries, as well
as roll combinations that lead to a result of zero.
* What this means is that, instead of trying to find an equation that
uses every die roll and matches a target number, the program can scan
for a smaller precalculated formula that uses some of the die rolls, and
add that to "zero" (made from the leftover rolls) to reach the same
result.
* This is dramatically faster over 99.9999% of the time, on the order of
a few microseconds (amortized) per run, for the miniscule tradeoff of
~29KB of space (less than the code takes) and ~100ms of boot time (not
noticeably slower even for single runs).
* Expressions are a lot more readable, both as a consequence of the
above and other cleanup and tweaks (adjusting order of operations,
removing unnecessary parenthesis, etc).
* The actual calculator bit is still available as a rarely-needed
fallback (and was used for the precalculation) but has been greatly
slimmed down, as the optimizations and complexity aren't much necessary.
Worst case tested peformance is still under a second.